Convert C_int and C_uint to take i64/u64 instead of host ints.

+++ src/librustc/back/abi.rs
@@ -75,7 +75,7 @@ pub static slice_elt_len: uint = 1u;
 
 pub static worst_case_glue_call_args: uint = 7u;
 
-pub static abi_version: uint = 1u;
+pub static abi_version: u64 = 1;
 
 pub fn memcpy_glue_name() -> ~str { return ~"rust_memcpy_glue"; }
